Split leaders into its own app #75
- Add leaders package
- Remove views from contributors
