+++ leaderboard/contributors/contributor_urls.py
@@ -1,0 +1,10 @@
+from django.conf.urls import url
+
+from leaderboard.contributors.views import (
+    UpdateContributorView,
+)
+
+urlpatterns = [
+    url('^(?P<uid>\w+)/$', UpdateContributorView.as_view(),
+        name='contributors-detail'),
+]
