+++ leaderboard/contributors/serializers.py
@@ -17,18 +17,6 @@ class ContributorUpdateSerializer(serializers.ModelSerializer):
         fields = ('name',)
 
 
-class ContributorSerializer(serializers.ModelSerializer):
-    """
-    Serialize a contributor with their name and the
-    number of observations they've made.
-    """
-    observations = serializers.IntegerField()
-
-    class Meta:
-        model = Contributor
-        fields = ('name', 'observations')
-
-
 class ContributionSerializer(serializers.Serializer):
     """
     A contribution submission from a contributor.  A contribution
